requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
dnspython==2.7.0
idna==3.7
pycryptodome==3.22.0
python-dateutil==2.9.0.post0
python-whois==0.9.5
requests==2.32.3
six==1.17.0
urllib3==2.2.2
